requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
xgboost>=2.0.0
joblib>=1.3.0
requests>=2.31.0
ta>=0.11.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
catboost>=1.2.2
python-dotenv>=1.0.0